+++ app/Http/Controllers/api/v1/RoomTypeController.php
@@ -98,6 +98,7 @@ class RoomTypeController extends Controller
         $roomType->max_participants = $request->max_participants;
         $roomType->max_duration = $request->max_duration;
         $roomType->create_parameters = $request->create_parameters;
+        $roomType->join_parameters = $request->join_parameters;
         $roomType->serverPool()->associate($request->server_pool);
 
         // Save default room settings
@@ -133,6 +134,7 @@ class RoomTypeController extends Controller
         $roomType->max_participants = $request->max_participants;
         $roomType->max_duration = $request->max_duration;
         $roomType->create_parameters = $request->create_parameters;
+        $roomType->join_parameters = $request->join_parameters;
         $roomType->serverPool()->associate($request->server_pool);
 
         // Safe default room settings