+++ lang/en/admin.php
@@ -153,6 +153,8 @@ return [
         'bbb_api' => [
             'create_parameters' => 'Additional Create API parameters',
             'create_parameters_description' => 'Specify as attribute-value pair (one per line, without spaces), e.g. webcamsOnlyForModerator=true',
+            'join_parameters' => 'Additional Join API parameters',
+            'join_parameters_description' => 'Specify as attribute-value pair (one per line, without spaces), e.g. userdata-bbb_show_session_details_on_join=false',
             'title' => 'BigBlueButton API',
         ],
         'color' => 'Color',