+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - User pictures to the admin user list ([#2131])
 - Last login datetime to the database ([#2150])
 - Last login datetime to the admin user list ([#2132], [#2150])
+- Custom join parameters in room type settings ([#2099], [#2151])
 
 ### Fixed
 
@@ -480,10 +481,12 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#2065]: https://github.com/THM-Health/PILOS/pull/2065
 [#2079]: https://github.com/THM-Health/PILOS/pull/2079
 [#2082]: https://github.com/THM-Health/PILOS/pull/2082
+[#2099]: https://github.com/THM-Health/PILOS/issues/2099
 [#2131]: https://github.com/THM-Health/PILOS/pull/2131
 [#2132]: https://github.com/THM-Health/PILOS/issues/2132
 [#2134]: https://github.com/THM-Health/PILOS/pull/2134
 [#2150]: https://github.com/THM-Health/PILOS/pull/2150
+[#2151]: https://github.com/THM-Health/PILOS/pull/2151
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v4.5.0...develop
 [v3.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.0
 [v3.0.1]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.1